ft: created a route function and its navigations
