+++ src/main.js
@@ -1,6 +1,10 @@
-import { createApp } from 'vue'
-import App from './App.vue'
+import { createApp } from "vue";
+import App from "./App.vue";
+import router from "./router/index";
 
-import './assets/main.css'
+import "./assets/main.css";
 
-createApp(App).mount('#app')
+const app = createApp(App);
+app.use(router);
+
+app.mount("#app");
