+++ src/views/QuizView.vue
@@ -1,0 +1,5 @@
+<template>
+  <div>
+    <h4>Hello quiz</h4>
+  </div>
+</template>
